+++ includes/class-remember.php
@@ -116,6 +116,7 @@ class Remember {
 		require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/utilities/class-remember-xmlrpc.php';
 		require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/utilities/class-remember-rest-privacy.php';
 		require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/utilities/class-remember-security-headers.php';
+		require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/utilities/class-remember-frontend.php';
 		require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/utilities/class-remember-logger.php';
 
 		$this->loader = new Remember_Loader();
@@ -217,6 +218,7 @@ class Remember {
 		Remember_Xmlrpc::init();
 		Remember_Rest_Privacy::init();
 		Remember_Security_Headers::init();
+		Remember_Frontend::init();
 	}
 
 	/**